requirements.txt
loguru
pip
python-dotenv
ruff
tqdm
typer
azure.storage.blob
pymongo
gspread
oauth2client
bcrypt
pandas
numpy
requests
datetime
streamlit
seaborn
ipykernel
neo4j~=5.28